Moved calculateIndexHoles out of DefaultQueryPlanner and other requested items.
