+++ warehouse/query-core/src/main/java/datawave/query/planner/FederatedQueryPlanner.java
@@ -1,12 +1,18 @@
 package datawave.query.planner;
 
+import com.google.common.collect.Multimap;
+import datawave.data.type.Type;
 import datawave.query.CloseableIterable;
 import datawave.query.config.FieldIndexHole;
 import datawave.query.config.ShardQueryConfiguration;
 import datawave.query.config.ValueIndexHole;
+import datawave.query.exceptions.DatawaveFatalQueryException;
 import datawave.query.exceptions.DatawaveQueryException;
+import datawave.query.exceptions.NoResultsException;
+import datawave.query.jexl.visitors.FetchDataTypesVisitor;
 import datawave.query.planner.pushdown.rules.PushDownRule;
 import datawave.query.tables.ScannerFactory;
+import datawave.query.util.IndexedDatesValue;
 import datawave.query.util.MetadataHelper;
 import datawave.query.util.YearMonthDay;
 import datawave.util.time.DateHelper;
@@ -14,6 +20,9 @@ import datawave.webservice.common.logging.ThreadConfigurableLogger;
 import datawave.webservice.query.Query;
 import datawave.webservice.query.configuration.GenericQueryConfiguration;
 import datawave.webservice.query.configuration.QueryData;
+import datawave.webservice.query.exception.DatawaveErrorCode;
+import datawave.webservice.query.exception.PreConditionFailedQueryException;
+import org.apache.accumulo.core.client.TableNotFoundException;
 import org.apache.accumulo.core.data.Key;
 import org.apache.accumulo.core.data.Value;
 import org.apache.accumulo.core.iterators.SortedKeyValueIterator;
@@ -35,26 +44,54 @@ public class FederatedQueryPlanner extends DefaultQueryPlanner {
         super();
     }
     
-    @Override
-    public FederatedQueryDataIterable process(GenericQueryConfiguration config, String query, Query settings, ScannerFactory scannerFactory)
+    public FederatedQueryDataIterable process2(GenericQueryConfiguration config, String query, Query settings, ScannerFactory scannerFactory)
                     throws DatawaveQueryException {
         
         FederatedQueryDataIterable returnQueryData = new FederatedQueryDataIterable();
         Date originalEndDate = config.getEndDate();
         Date originalStartDate = config.getBeginDate();
         TreeSet<YearMonthDay> holeDates;
+        MetadataHelper metadataHelper = getMetadataHelper();
         
         // The DefaultQueryPlanner.process needs to be called so that the FieldIndexHoles can be calculated
         // and the queryData returned will be used if there are no index holes of any kind.
-        CloseableIterable<QueryData> queryData = super.process(config, query, settings, scannerFactory);
+        final QueryData queryData = new QueryData();
+        CloseableIterable<QueryData> results;
         
-        setDoCalculateFieldIndexHoles(false);
+        results = super.process(config, query, settings, scannerFactory);
         
         if (config instanceof ShardQueryConfiguration) {
+            ASTJexlScript queryTree = null;
+            try {
+                queryTree = updateQueryTree(scannerFactory, metadataHelper, dateIndexHelper, (ShardQueryConfiguration) config, query, queryData, settings);
+            } catch (StackOverflowError e) {
+                if (log.isTraceEnabled()) {
+                    log.trace("Stack trace for overflow " + e);
+                }
+                PreConditionFailedQueryException qe = new PreConditionFailedQueryException(DatawaveErrorCode.QUERY_DEPTH_OR_TERM_THRESHOLD_EXCEEDED, e);
+                log.warn(qe);
+                throw new DatawaveFatalQueryException(qe);
+            } catch (NoResultsException e) {
+                if (log.isTraceEnabled()) {
+                    log.trace("Definitively determined that no results exist from the indexes");
+                }
+                
+            }
+            
+            Multimap<String,Type<?>> fieldToDatatypeMap = FetchDataTypesVisitor.fetchDataTypes(metadataHelper,
+                            ((ShardQueryConfiguration) config).getDatatypeFilter(), queryTree, false);
+            
+            try {
+                
+                calculateFieldIndexHoles(metadataHelper, fieldToDatatypeMap, (ShardQueryConfiguration) config);
+            } catch (TableNotFoundException e) {
+                log.error("metadata table was not found " + e.getMessage());
+            }
+            
             List<FieldIndexHole> fieldIndexHoles = ((ShardQueryConfiguration) config).getFieldIndexHoles();
             List<ValueIndexHole> valueIndexHoles = ((ShardQueryConfiguration) config).getValueIndexHoles();
             if ((valueIndexHoles == null && fieldIndexHoles == null) || (valueIndexHoles.size() == 0 && fieldIndexHoles.size() == 0)) {
-                returnQueryData.addDelegate(queryData);
+                returnQueryData.addDelegate(results);
                 return returnQueryData;
             }
             
@@ -79,15 +116,15 @@ public class FederatedQueryPlanner extends DefaultQueryPlanner {
                     }
                 }
                 
-                queryData = getQueryData((ShardQueryConfiguration) config, query, settings, scannerFactory, startDate, endDate);
-                returnQueryData.addDelegate(queryData);
+                results = getQueryData((ShardQueryConfiguration) config, query, settings, scannerFactory, startDate, endDate);
+                returnQueryData.addDelegate(results);
                 
             }
             
         }
         
         if (!returnQueryData.iterator().hasNext())
-            returnQueryData.addDelegate(queryData);
+            returnQueryData.addDelegate(results);
         
         return returnQueryData;
     }
@@ -98,8 +135,44 @@ public class FederatedQueryPlanner extends DefaultQueryPlanner {
         ShardQueryConfiguration tempConfig = new ShardQueryConfiguration(config);
         tempConfig.setBeginDate(startDate);
         tempConfig.setEndDate(endDate);
-        queryData = super.process(tempConfig, query, settings, scannerFactory);
+        queryData = process2(tempConfig, query, settings, scannerFactory);
         return queryData;
+    }
+    
+    private TreeSet<YearMonthDay> generateStartAndEndDates(ShardQueryConfiguration configuration) {
+        
+        String startDate = DateHelper.format(configuration.getBeginDate().getTime());
+        String endDate = DateHelper.format(configuration.getEndDate().getTime());
+        
+        YearMonthDay.Bounds bounds = new YearMonthDay.Bounds(startDate, true, endDate, true);
+        
+        TreeSet<YearMonthDay> queryDates = new TreeSet<>();
+        for (ValueIndexHole valueIndexHole : configuration.getValueIndexHoles()) {
+            addDatesToSet(bounds, queryDates, valueIndexHole.getStartDate());
+            addDatesToSet(bounds, queryDates, valueIndexHole.getEndDate());
+        }
+        
+        for (FieldIndexHole fieldIndexHole : configuration.getFieldIndexHoles()) {
+            // TODO remove comparison below. calculateFieldHoles needs to be fixed.
+            if (fieldIndexHole.getStartDate().compareTo(fieldIndexHole.getEndDate()) <= 0) {
+                addDatesToSet(bounds, queryDates, fieldIndexHole.getStartDate());
+                addDatesToSet(bounds, queryDates, fieldIndexHole.getEndDate());
+            }
+        }
+        
+        return queryDates;
+    }
+    
+    private void addDatesToSet(YearMonthDay.Bounds bounds, TreeSet<YearMonthDay> queryDates, String strDate) {
+        if (bounds.withinBounds(strDate))
+            queryDates.add(new YearMonthDay(strDate));
+    }
+    
+    @Override
+    public FederatedQueryDataIterable process(GenericQueryConfiguration config, String query, Query settings, ScannerFactory scannerFactory)
+                    throws DatawaveQueryException {
+        
+        return process2(config, query, settings, scannerFactory);
     }
     
     @Override
@@ -147,33 +220,116 @@ public class FederatedQueryPlanner extends DefaultQueryPlanner {
         return super.applyRules(queryTree, scannerFactory, metadataHelper, config);
     }
     
-    private TreeSet<YearMonthDay> generateStartAndEndDates(ShardQueryConfiguration configuration) {
+    /**
+     * Calculate the FieldIndexHoles and add them to the ShardedQueryConfiguaration
+     *
+     * @param metadataHelper
+     * @param fieldToDatatypeMap
+     * @param config
+     */
+    public void calculateFieldIndexHoles(MetadataHelper metadataHelper, Multimap<String,Type<?>> fieldToDatatypeMap, ShardQueryConfiguration config)
+                    throws TableNotFoundException {
         
-        String startDate = DateHelper.format(configuration.getBeginDate().getTime());
-        String endDate = DateHelper.format(configuration.getEndDate().getTime());
-        
+        IndexedDatesValue indexedDates;
+        String startDate = DateHelper.format(config.getBeginDate().getTime());
+        String endDate = DateHelper.format(config.getEndDate().getTime());
+        String holeStart = startDate;
+        String lastHoleEndate = null;
         YearMonthDay.Bounds bounds = new YearMonthDay.Bounds(startDate, true, endDate, true);
+        FieldIndexHole newHole = null;
+        boolean firstHole = true;
+        boolean foundHolesInDateBounds = false;
+        String previousDay, nextDay = null;
+        log.debug("startDate is: " + startDate + " and endDate is " + endDate);
         
-        TreeSet<YearMonthDay> queryDates = new TreeSet<>();
-        for (ValueIndexHole valueIndexHole : configuration.getValueIndexHoles()) {
-            addDatesToSet(bounds, queryDates, valueIndexHole.getStartDate());
-            addDatesToSet(bounds, queryDates, valueIndexHole.getEndDate());
-        }
-        
-        for (FieldIndexHole fieldIndexHole : configuration.getFieldIndexHoles()) {
-            // TODO remove comparison below. calculateFieldHoles needs to be fixed.
-            if (fieldIndexHole.getStartDate().compareTo(fieldIndexHole.getEndDate()) <= 0) {
-                addDatesToSet(bounds, queryDates, fieldIndexHole.getStartDate());
-                addDatesToSet(bounds, queryDates, fieldIndexHole.getEndDate());
+        for (String field : fieldToDatatypeMap.keySet()) {
+            indexedDates = metadataHelper.getIndexDates(field, config.getDatatypeFilter());
+            if (indexedDates != null && !(indexedDates.getIndexedDatesSet().size() == 0)) {
+                for (YearMonthDay entry : indexedDates.getIndexedDatesSet()) {
+                    // Only create a hole if the indexed field are within date bounds
+                    if (bounds.withinBounds(entry)) {
+                        foundHolesInDateBounds = true;
+                        if (firstHole && holeStart.compareTo(entry.getYyyymmdd()) < 0) {
+                            // create the FieldIndexHole for the dates the field was not indexed before the first
+                            // time in the date range that it was indexed.
+                            FieldIndexHole firstIndexHole = new FieldIndexHole(field, startDate);
+                            previousDay = previousDay(entry.getYyyymmdd());
+                            nextDay = nextDay(entry.getYyyymmdd());
+                            log.debug("The date in the entry is: " + entry.getYyyymmdd());
+                            log.debug("The previous day is: " + previousDay);
+                            log.debug("The next day is: " + nextDay);
+                            firstIndexHole.setEndDate(previousDay);
+                            config.addFieldIndexHole(firstIndexHole);
+                            holeStart = nextDay(entry.getYyyymmdd());
+                            firstHole = false;
+                        }
+                        
+                        // The end date of the last hole processed depends on the next date the field was indexed
+                        if (newHole != null) {
+                            lastHoleEndate = previousDay(entry.getYyyymmdd());
+                            newHole.setEndDate(lastHoleEndate);
+                        } else {
+                            lastHoleEndate = nextDay(entry.getYyyymmdd());
+                            if (lastHoleEndate.compareTo(endDate) > 0)
+                                lastHoleEndate = endDate;
+                        }
+                        
+                        /*
+                         * If the start of the next potential hole is the same as the date indexed field there is no need to start creating and index hole
+                         * starting on that date so increment the holeStart and find the next indexed date.
+                         */
+                        if (holeStart.equals(entry.getYyyymmdd())) {
+                            holeStart = nextDay(entry.getYyyymmdd());
+                            continue;
+                        }
+                        
+                        // At this point, create a new FieldIndexHole
+                        if (bounds.withinBounds(holeStart)) {
+                            newHole = new FieldIndexHole();
+                            newHole.setFieldName(field);
+                            newHole.setStartDate(holeStart);
+                            // you have to see next date the the field was indexed in the next iteration
+                            // before you can set the end date. That may get done outside the loop on line 1698
+                            // or on 1669 with the previous day of the next date the field was indexed
+                            config.addFieldIndexHole(newHole);
+                        }
+                    }
+                    holeStart = nextDay(entry.getYyyymmdd());
+                }
             }
+            
+            if (newHole != null)
+                newHole.setEndDate(lastHoleEndate);
+            
+            if (foundHolesInDateBounds && bounds.withinBounds(holeStart)) {
+                FieldIndexHole trailingHole = new FieldIndexHole(field, new String[] {holeStart, endDate});
+                config.addFieldIndexHole(trailingHole);
+            }
+            
+            if (!foundHolesInDateBounds) {
+                FieldIndexHole trailingHole = new FieldIndexHole(field, new String[] {startDate, endDate});
+                config.addFieldIndexHole(trailingHole);
+            }
+            
+            if (!config.getFieldIndexHoles().isEmpty()) {
+                log.debug("Found Field index holes for field: " + field + " within date bounds");
+                for (FieldIndexHole hole : config.getFieldIndexHoles()) {
+                    log.debug(hole.toString());
+                }
+            } else {
+                log.debug("No fieldHoles created.");
+            }
+            
         }
         
-        return queryDates;
     }
     
-    private void addDatesToSet(YearMonthDay.Bounds bounds, TreeSet<YearMonthDay> queryDates, String strDate) {
-        if (bounds.withinBounds(strDate))
-            queryDates.add(new YearMonthDay(strDate));
+    private static String previousDay(String day) {
+        return YearMonthDay.previousDay(day).getYyyymmdd();
+    }
+    
+    private static String nextDay(String day) {
+        return YearMonthDay.nextDay(day).getYyyymmdd();
     }
     
 }
